move file (logger)
